+++ src/main/java/io/github/tehstoneman/redstonepp/ModInfo.java
@@ -1,6 +1,0 @@
-package io.github.tehstoneman.redstonepp;
-
-public class ModInfo
-{
-	public static final String MOD_ID = "redstonepp";
-}