+++ src/main/java/io/github/tehstoneman/redstonextra/ModInfo.java
@@ -1,0 +1,6 @@
+package io.github.tehstoneman.redstonextra;
+
+public class ModInfo
+{
+	public static final String MOD_ID = "redstonextra";
+}