+++ CHANGELOG.md
@@ -10,3 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Removed
+
+## [1.0.0.0_beta1] - 2020-07-25
+### Added
+- And gate
+
+### Changed
+- Renamed project to RedstoneXtra
+
+## [1.0.0.0_beta0] - 2020-07-11
+### Added
+- Redstone Inverter
+- Initial release